Lesson 23, step 4, size_t usage

+++ 23-fixes/libc/mem.c
@@ -18,7 +18,7 @@ void memory_set(uint8_t *dest, uint8_t val, uint32_t len) {
 uint32_t free_mem_addr = 0x10000;
 /* Implementation is just a pointer to some free memory which
  * keeps growing */
-uint32_t kmalloc(uint32_t size, int align, uint32_t *phys_addr) {
+uint32_t kmalloc(size_t size, int align, uint32_t *phys_addr) {
     /* Pages are aligned to 4K, or 0x1000 */
     if (align == 1 && (free_mem_addr & 0xFFFFF000)) {
         free_mem_addr &= 0xFFFFF000;
